Improved code in Game section, reduced delay for impproved blast checking
