+++ src/org/academiadecodigo/bootcamp/Location.java
@@ -10,8 +10,8 @@ public class Location {
 
     public Location(Map map) {
         this.map = map;
-        cols = (int) Math.floor(map.getWidth() * Math.random())+1;
-        rows = (int) Math.floor(map.getHeight() * Math.random())+1;
+        cols = (int) Math.floor((map.getWidth()-1) * Math.random())+1;
+        rows = (int) Math.floor((map.getHeight()-1) * Math.random())+1;
     }
 
     public Location(int cols, int rows, Map map){
