+++ src/org/academiadecodigo/bootcamp/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static void main(String[] args) throws InterruptedException {
 
-        Game g = new Game( 200);
+        Game g = new Game( 30);
 
         g.init();
         g.start();
